fixup/key_scanner: shorten delay for `read_matrix`
Shortens the delay in the inner loop of the key scanner to improve
responsiveness without spamming repeated keys.

Setting the delay too low results in many repeated keys being registered
on a single key press. Setting the value too high makes the keyboard
feel sticky and unresponsive.

The current middle value seems to be a nice balance. There may end up
being a cleaner solution for slowing down the processor here, but this
works for now.