key_scanner/usb_context: increase safety of globals
Switches the flag for performing a key scan into a global atomic
variable. This removes the need to access the flag through the
`UsbContext`.

Wraps the global `UsbContext` in a `Mutex<RefCell<Option<UsbContext>>>`,
which allows interior mutability, and removes the need to declare
`USB_CONTEXT` as `static mut`.

Increases the safety of `USB_CONTEXT` access by requiring a lock to be
acquired. Not incredibly relevant currently, since `USB_CONTEXT` is
only accessed inside two interrupt handlers, and only one appears to be
firing (`USB_GEN`). However, it is good for future-proofing.